+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Changes to be tagged
 ### Added
+- `mstSvtIndividuality` to raw svt and `traitAdd` to nice svt.
 - charaGraphEx to servant assets.
 - skillAdd to skill.